Rename attr to xattr
