+++ src/xattr/mod.rs
@@ -1,10 +1,10 @@
 //! Extended attribute support
 #[cfg(target_os = "macos")]
-mod attr_darwin;
+mod xattr_darwin;
 #[cfg(target_os = "macos")]
-pub use self::attr_darwin::*;
+pub use self::xattr_darwin::*;
 #[cfg(not(target_os = "macos"))]
-mod attr_other;
+mod xattr_other;
 #[cfg(not(target_os = "macos"))]
-pub use self::attr_other::*;
+pub use self::xattr_other::*;
